+++ lib/core/evaluation.sh
@@ -13,6 +13,7 @@ shellspec_syntax 'shellspec_evaluation_run'
 shellspec_proxy 'shellspec_evaluation' 'shellspec_syntax_dispatch evaluation'
 
 shellspec_evaluation_call() {
+  shellspec_coverage_start
   if [ ! "${SHELLSPEC_DATA:-}" ]; then
     shellspec_around_call "$@" < "$SHELLSPEC_STDIN_DEV"
   else
@@ -20,14 +21,17 @@ shellspec_evaluation_call() {
     shellspec_around_call "$@" < "$SHELLSPEC_STDIN_FILE"
   fi >"$SHELLSPEC_STDOUT_FILE" 2>"$SHELLSPEC_STDERR_FILE" &&:
   shellspec_evaluation_cleanup $?
+  shellspec_coverage_stop
 }
 
 shellspec_evaluation_run() {
+  shellspec_coverage_start
   case $- in
     *e*) set +e; shellspec_evaluation_run_subshell -e "$@"; set -e -- $? ;;
     *) shellspec_evaluation_run_subshell +e "$@"; set -- $? ;;
   esac
   shellspec_evaluation_cleanup "$1"
+  shellspec_coverage_stop
 }
 
 if [ "${SHELLSPEC_SHELL_TYPE#p}" = "bosh" ]; then
@@ -45,11 +49,13 @@ fi
 
 shellspec_evaluation_run_data() {
   if [ ! "${SHELLSPEC_DATA:-}" ]; then
-    shellspec_evaluation_run_trap_exit_status "$@" < "$SHELLSPEC_STDIN_DEV"
+    shellspec_evaluation_run_trap_exit_status "$@" < "$SHELLSPEC_STDIN_DEV" \
+      >"$SHELLSPEC_STDOUT_FILE" 2>"$SHELLSPEC_STDERR_FILE"
   else
     shellspec_data > "$SHELLSPEC_STDIN_FILE"
-    shellspec_evaluation_run_trap_exit_status "$@" < "$SHELLSPEC_STDIN_FILE"
-  fi >"$SHELLSPEC_STDOUT_FILE" 2>"$SHELLSPEC_STDERR_FILE"
+    shellspec_evaluation_run_trap_exit_status "$@" < "$SHELLSPEC_STDIN_FILE" \
+      >"$SHELLSPEC_STDOUT_FILE" 2>"$SHELLSPEC_STDERR_FILE"
+  fi
 }
 
 if [ "${ZSH_VERSION:-}" ] && (exit 1); then
@@ -86,10 +92,47 @@ fi
 
 shellspec_evaluation_run_instruction() {
   case $1 in
+    script) shift; shellspec_evaluation_run_script "$@" ;;
     command) shift; shellspec_evaluation_run_command "$@" ;;
     source) shift; shellspec_evaluation_run_source "$@" ;;
     *) "$@" ;;
   esac
+}
+
+shellspec_shebang_arguments() {
+  read -r line
+  case $line in (\#!/usr/bin/env\ * | \#!/bin/env\ *)
+    shellspec_trim line "${line#* }"
+    line="#!$line"
+  esac
+  case $line in (\#!*)
+    shellspec_trim line "$line"
+    case $line in (*\ *)
+      shellspec_trim line "${line#* }"
+      shellspec_putsn "$line"
+    esac
+  esac
+}
+
+shellspec_evaluation_run_script() {
+  if [ ! -e "$1" ]; then
+    eval "$SHELLSPEC_SHELL \"\$@\""
+  elif [ ! -x "$1" ]; then
+    command "$@"
+  else
+    if [ "$SHELLSPEC_SHEBANG_MULTIARG" ]; then
+      IFS=" $IFS"
+      # shellcheck disable=SC2046
+      set -- $(shellspec_shebang_arguments < "$1") "$@"
+      IFS=${IFS# }
+    else
+      set -- "$(shellspec_shebang_arguments < "$1")" "$@"
+      [ "$1" ] || shift
+    fi
+    # shellcheck disable=SC1090
+    [ "$SHELLSPEC_COVERAGE_ENV" ] && . "$SHELLSPEC_COVERAGE_ENV"
+    eval "$SHELLSPEC_SHELL \"\$@\""
+  fi
 }
 
 shellspec_evaluation_run_command() {