+++ lib/cov/kcov-env.sh
@@ -1,0 +1,3 @@
+#shellcheck shell=bash
+
+export BASH_ENV="$SHELLSPEC_TMPBASE/kcov-debug-helper.sh"